requirements.txt
python-telegram-bot==20.7
reportlab==4.0.4